requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.0.247
langchain-community
openai>=0.27.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.4
pypdf>=3.8.1
sentence-transformers>=2.2.2
tiktoken